requirements.txt
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
easydict
Flask==1.1.4
Flask-CLI==0.4.0
Flask-Cors==3.0.10
gunicorn==20.1.0
idna==2.10
importlib-metadata==2.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.19.2
packaging==20.9
pandas
pip==20.3.4
pretty_midi
pyparsing==2.4.7
python-dotenv==0.19.1
pytz==2021.1
requests==2.25.1
simpletransformers
six~=1.15.0
tensorflow==2.5.0
torch
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==1.0.1
zipp==1.2.0
